removes WebsiteComposer, BackOfficeComposer
