+++ src/Umbraco.Web.BackOffice/DependencyInjection/UmbracoBuilderExtensions.cs
@@ -1,11 +1,21 @@
 using System;
+using System.Linq;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging;
 using Umbraco.Core.DependencyInjection;
+using Umbraco.Core.Hosting;
+using Umbraco.Core.IO;
+using Umbraco.Core.Services;
+using Umbraco.Extensions;
 using Umbraco.Infrastructure.DependencyInjection;
-using Umbraco.Infrastructure.PublishedCache.DependencyInjection;
 using Umbraco.Web.BackOffice.Authorization;
+using Umbraco.Web.BackOffice.Controllers;
+using Umbraco.Web.BackOffice.Filters;
+using Umbraco.Web.BackOffice.Middleware;
+using Umbraco.Web.BackOffice.Routing;
 using Umbraco.Web.BackOffice.Security;
+using Umbraco.Web.BackOffice.Services;
 using Umbraco.Web.BackOffice.Trees;
 using Umbraco.Web.Common.Authorization;
 using Umbraco.Web.Common.DependencyInjection;
@@ -25,6 +35,7 @@ namespace Umbraco.Web.BackOffice.DependencyInjection
                 .AddUmbracoCore()
                 .AddWebComponents()
                 .AddRuntimeMinifier()
+                .AddBackOfficeCore()
                 .AddBackOfficeAuthentication()
                 .AddBackOfficeIdentity()
                 .AddBackOfficeAuthorizationPolicies()
@@ -64,6 +75,11 @@ namespace Umbraco.Web.BackOffice.DependencyInjection
                 });
 
             builder.Services.ConfigureOptions<ConfigureBackOfficeCookieOptions>();
+
+            builder.Services.AddUnique<PreviewAuthenticationMiddleware>();
+            builder.Services.AddUnique<BackOfficeExternalLoginProviderErrorMiddleware>();
+            builder.Services.AddUnique<IBackOfficeAntiforgery, BackOfficeAntiforgery>();
+
             return builder;
         }
 
@@ -117,5 +133,41 @@ namespace Umbraco.Web.BackOffice.DependencyInjection
         /// </summary>
         public static TreeCollectionBuilder Trees(this IUmbracoBuilder builder)
             => builder.WithCollectionBuilder<TreeCollectionBuilder>();
+
+        public static IUmbracoBuilder AddBackOfficeCore(this IUmbracoBuilder builder)
+        {
+            builder.Services.AddUnique<BackOfficeAreaRoutes>();
+            builder.Services.AddUnique<PreviewRoutes>();
+            builder.Services.AddUnique<BackOfficeServerVariables>();
+            builder.Services.AddScoped<BackOfficeSessionIdValidator>();
+            builder.Services.AddScoped<BackOfficeSecurityStampValidator>();
+
+            // register back office trees
+            // the collection builder only accepts types inheriting from TreeControllerBase
+            // and will filter out those that are not attributed with TreeAttribute
+            var umbracoApiControllerTypes = builder.TypeLoader.GetUmbracoApiControllers().ToList();
+            builder.Trees()
+                .AddTreeControllers(umbracoApiControllerTypes.Where(x => typeof(TreeControllerBase).IsAssignableFrom(x)));
+
+            builder.ComposeWebMappingProfiles();
+
+            builder.Services.AddUnique<IPhysicalFileSystem>(factory =>
+            {
+                var path = "~/";
+                var hostingEnvironment = factory.GetRequiredService<IHostingEnvironment>();
+                return new PhysicalFileSystem(
+                    factory.GetRequiredService<IIOHelper>(),
+                    hostingEnvironment,
+                    factory.GetRequiredService<ILogger<PhysicalFileSystem>>(),
+                    hostingEnvironment.MapPathContentRoot(path),
+                    hostingEnvironment.ToAbsolute(path)
+                );
+            });
+
+            builder.Services.AddUnique<IIconService, IconService>();
+            builder.Services.AddUnique<UnhandledExceptionLoggerMiddleware>();
+
+            return builder;
+        }
     }
 }
