+++ src/Umbraco.Web.UI.NetCore/Startup.cs
@@ -46,6 +46,7 @@ namespace Umbraco.Web.UI.NetCore
             services.AddUmbraco(_env, _config)
                 .AddBackOffice()
                 .AddWebsite()
+                .AddComposers()
                 .Build();
 #pragma warning restore IDE0022 // Use expression body for methods
