+++ src/Umbraco.Web.Website/DependencyInjection/UmbracoBuilderExtensions.cs
@@ -2,8 +2,10 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
 using Umbraco.Core.DependencyInjection;
+using Umbraco.Extensions;
 using Umbraco.Infrastructure.DependencyInjection;
 using Umbraco.Infrastructure.PublishedCache.DependencyInjection;
+using Umbraco.Web.Website.Collections;
 using Umbraco.Web.Website.Controllers;
 using Umbraco.Web.Website.Routing;
 using Umbraco.Web.Website.ViewEngines;
@@ -20,6 +22,11 @@ namespace Umbraco.Web.Website.DependencyInjection
         /// </summary>
         public static IUmbracoBuilder AddWebsite(this IUmbracoBuilder builder)
         {
+            builder.Services.AddUnique<NoContentRoutes>();
+
+            builder.WithCollectionBuilder<SurfaceControllerTypeCollectionBuilder>()
+                 .Add(builder.TypeLoader.GetSurfaceControllers());
+
             // Set the render & plugin view engines (Super complicated, but this allows us to use the IServiceCollection
             // to inject dependencies into the viewEngines)
             builder.Services.AddTransient<IConfigureOptions<MvcViewOptions>, RenderMvcViewOptionsSetup>();
